fix: preventing html escaping + module renaming
