+++ disma-cli/Cargo.toml
@@ -18,7 +18,7 @@ path = "src/main.rs"
 [dependencies]
 disma = { path = "../disma", version = "0.16.2" }
 clap = { version = "4.4.16", features = ["derive"] }
-handlebars = "4.3.3"
+handlebars = "5.1.1"
 serde = { version = "1.0.143", features = ["derive"] }
 serde_json = "1.0.83"
 serde_yaml = "0.9.9"
