+++ disma-cli/src/commands/list_guilds.rs
