+++ disma-cli/src/main.rs
@@ -1,8 +1,8 @@
 mod actions;
 mod args;
+mod commands;
 mod infra;
 mod injector;
-mod services;
 mod utils;
 
 use clap::Parser;
